app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
azure-core==1.35.0
azure-identity==1.23.1
azure-monitor-opentelemetry-exporter==1.0.0b40
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.2.1
cryptography==45.0.5
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.116.1
firebase_admin==7.0.0
fixedint==0.1.6
fonttools==4.53.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
numpy==2.0.0
oauthlib==3.3.1
opentelemetry-api==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-requests==0.56b0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.2
zipp==3.23.0
